+++ src/blocking.rs
@@ -491,7 +491,8 @@ impl KiCadClientBlocking {
         fn get_items_by_net_class(&self, type_codes: Vec<i32>, net_classes: Vec<String>) -> Result<Vec<PcbItem>, KiCadError>;
         fn get_connected_items_raw(&self, item_ids: Vec<String>, type_codes: Vec<i32>) -> Result<Vec<Any>, KiCadError>;
         fn get_connected_items(&self, item_ids: Vec<String>, type_codes: Vec<i32>) -> Result<Vec<PcbItem>, KiCadError>;
-        fn get_netclass_for_nets_raw(&self, nets: Vec<BoardNet>) -> Result<Any, KiCadError>;        fn get_netclass_for_nets(&self, nets: Vec<BoardNet>) -> Result<Vec<NetClassForNetEntry>, KiCadError>;
+        fn get_netclass_for_nets_raw(&self, nets: Vec<BoardNet>) -> Result<Any, KiCadError>;
+        fn get_netclass_for_nets(&self, nets: Vec<BoardNet>) -> Result<Vec<NetClassForNetEntry>, KiCadError>;
         fn refill_zones(&self, zone_ids: Vec<String>) -> Result<(), KiCadError>;
         fn refill_all_zones(&self) -> Result<(), KiCadError>;
         fn get_pad_shape_as_polygon_raw(&self, pad_ids: Vec<String>, layer_id: i32) -> Result<Vec<Any>, KiCadError>;
@@ -512,7 +513,8 @@ impl KiCadClientBlocking {
         fn get_title_block_info(&self) -> Result<TitleBlockInfo, KiCadError>;
         fn set_title_block_info_raw(&self, title_block: TitleBlockInfo) -> Result<Any, KiCadError>;
         fn set_title_block_info(&self, title_block: TitleBlockInfo) -> Result<(), KiCadError>;
-        fn save_document_raw(&self) -> Result<Any, KiCadError>;        fn save_document(&self) -> Result<(), KiCadError>;
+        fn save_document_raw(&self) -> Result<Any, KiCadError>;
+        fn save_document(&self) -> Result<(), KiCadError>;
         fn revert_document_raw(&self) -> Result<Any, KiCadError>;
         fn revert_document(&self) -> Result<(), KiCadError>;
         fn get_board_as_string(&self) -> Result<String, KiCadError>;